avances Se agrega configuracion para loggin, container de DI y mapper Se agrega interfaces de loggin para DI Se agregan nuevas clases de business y se agregan nuevas funcionalidades a la base Se agregan Filters/UsuarioFilter.csSe prepara el bootstraper para ser llamado desde el main form Program.cs

+++ PalcoNet.Dtos/Filters/UsuarioFilter.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace PalcoNet.Infraestructure.Filters
 {
-    public class UsuarioFilter : BaseFilter<int>
+    public class UsuarioFilter : BaseFilter<Int64>
     {
     }
 }
